Adds capability to change RF Bandwidth in Pluto FFT.
Pluto FFT script had a fixed RF Bandwidth of operation (20 MHz).
This commit adds a parameter that can be used to dynamically tune
the RF bandwidth while visualizing the spectrum. The ability to
change the RF bandwidth and the baseband sample rate together
can enable better signal selectivity and visualization.